requirements-3.8.6.txt
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
idna==2.10
lxml==4.6.1
Pillow==8.0.0
pytesseract==0.3.6
python-docx==0.8.10
requests==2.24.0
selenium==3.141.0
urllib3==1.25.11
webdriver-manager==3.2.2
